# optimized.js 12% faster using {num:1} and readSync