Based on the discussion in https://github.com/benhoyt/countwords/pull/71:

Using `{num:1}` and `fs.readSync(process.stdin.fd)` I got ~12% faster:

Current optimized.js:
```
    Benchmark #1: node ./optimized.js < ./kjvbible_x10.txt > /dev/null
       Time (mean ± σ):      1.006 s ±  0.017 s    [User: 918.4 ms, System: 126.3 ms]
       Range (min … max):    0.982 s …  1.086 s    100 runs
```

With `new RefNum(1)`
```
    Benchmark #1: node ./optimizedRefNum.js < ./kjvbible_x10.txt > /dev/null
       Time (mean ± σ):     969.5 ms ±  36.9 ms    [User: 888.0 ms, System: 127.5 ms]
       Range (min … max):   934.3 ms … 1091.8 ms    100 runs
```

With `{num:1}`
```
    Benchmark #1: node ./optimizedObj.js < ./kjvbible_x10.txt > /dev/null
        Time (mean ± σ):     954.0 ms ±  10.7 ms    [User: 874.1 ms, System: 126.0 ms]
        Range (min … max):   938.7 ms … 989.9 ms    100 runs
```

With `fs.readSync`
```
    Benchmark #1: node ./optimizedReadSync.js < ./kjvbible_x10.txt > /dev/null
       Time (mean ± σ):     963.7 ms ±  24.4 ms    [User: 883.0 ms, System: 114.6 ms]
       Range (min … max):   941.1 ms … 1121.1 ms    100 runs
```

With `{num:1}` and `fs.readSync`
```
    Benchmark #1: node ./optimizedObjReadSync.js < ./kjvbible_x10.txt > /dev/null
        Time (mean ± σ):     886.2 ms ±  13.6 ms    [User: 807.8 ms, System: 114.2 ms]
        Range (min … max):   868.0 ms … 943.2 ms    100 runs
```