@benhoyt 2021-03-18
Thanks for this. Yeah, it's maybe 5% faster for me. Not quite sure it's worth it, but I'll merge and update the results later!