@carstenschwede 2021-03-18
Interesting. Are you using node.js >= 14? Which platform? Maybe the minimum out of 5 iterations in your benchmark is not sufficient to account for warmup effects? In the 100 runs on macOS/node v14, JS always significantly outperforms Python3 in both the simple and optimized versions (i.e. the slowest JS run is always faster than the fasted Python3 run). Might be worth a disclaimer on the results page.

Python3

```
Benchmark #1: python3 simple.py < ./kjvbible_x10.txt > /dev/null
  Time (mean ± σ):      2.551 s ±  0.148 s    [User: 2.457 s, System: 0.121 s]
  Range (min … max):    2.344 s …  3.353 s    100 runs
```

```
Benchmark #1: python3 optimized.py < ./kjvbible_x10.txt > /dev/null
  Time (mean ± σ):      1.528 s ±  0.063 s    [User: 1.440 s, System: 0.116 s]
  Range (min … max):    1.466 s …  1.877 s    100 runs
```

******

JavaScript:

```
Benchmark #1: node ./simple.js < ./kjvbible_x10.txt > /dev/null
  Time (mean ± σ):      1.359 s ±  0.096 s    [User: 1.266 s, System: 0.134 s]
  Range (min … max):    1.287 s …  1.678 s    100 runs
```

```
Benchmark #1: node ./optimizedObjReadSync.js < ./kjvbible_x10.txt > /dev/null
        Time (mean ± σ):     886.2 ms ±  13.6 ms    [User: 807.8 ms, System: 114.2 ms]
        Range (min … max):   868.0 ms … 943.2 ms    100 runs
```